chore: 🔥 add real data and validations
- Add data for inspection screens
- Rename a screen widget
- Change data random defaults to real data
- Change screen width to media query width
- Adding the Inspection Progress Time property
- Add Inspection Cost Property

+++ lib/data/models/detail_image.dart
@@ -12,17 +12,17 @@ class DetailImage extends Equatable {
     return _$DetailImageFromJson(json);
   }
 
-  final double? height;
+  final int? height;
   @JsonKey(name: 'image_path')
   final String? imagePath;
-  final double? width;
+  final int? width;
 
   Map<String, dynamic> toJson() => _$DetailImageToJson(this);
 
   DetailImage copyWith({
-    double? height,
+    int? height,
     String? imagePath,
-    double? width,
+    int? width,
   }) {
     return DetailImage(
       height: height ?? this.height,
